+++ server/globals.js
@@ -6,15 +6,11 @@ var appRoot = path.join(__dirname, '..');
 global.appRoot = appRoot;
 
 function restrict (req, res, next) {
-    if (global.authentication) {
-        if (req.isAuthenticated()) {
-            next();
-        } else {
-            req.session.error = 'Access denied!';
-            return res.redirect(302, '/login');
-        }
-    } else {
+    if (req.isAuthenticated()) {
         next();
+    } else {
+        req.session.error = 'Access denied!';
+        return res.status(403).send('Access denied');
     }
 }
 global.restrict = restrict;